rename EditorText to Editor
