+++ catalog/app/components/Preview/renderers/Editor.tsx
