+++ catalog/app/components/Preview/renderers/index.js
@@ -1,7 +1,7 @@
 export { default as Audio } from './Audio'
 export { default as DataFrame } from './DataFrame'
 export { default as ECharts } from './ECharts'
-export { default as EditorText } from './EditorText'
+export { default as Editor } from './Editor'
 export { default as Fcs } from './Fcs'
 export { default as IFrame } from './IFrame'
 export { default as Image } from './Image'
